requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
openai>=1.0.0
google-genai>=0.3.0
